Add sample input for command line execution.
The section now has two sample input. One is for the call to
execute the program from the command line with the name of the
file containing the text as an argument. The other is a small
amount of text from the sample input given in this example.

Files have been renamed as requested and appropriate changes have
been made to the related files.