Fix another issue with KMP implementation.
When looking at cases where matches were adjacent to each other,
the algorithm would fail to detect the second valid match. This
was caused by the calculation of the length of the obtained
partial match. Prior, it was always increased by one, when, in
fact, the actual length of the partial match was simply 'm'.