Add test input file.
This file contains test cases for the algorithm. The number of
test cases is easily extendable by simply adding an empty line
followed by three more lines containing: the expected number of
matches, the string to search for (can include spaces) and also
the string to search.

The testing is done so that one line is used, but that could
change, all depending on how text is to be read in. The algorithm
will work the same.